Add an overlay layer to the viewport.
Within the overlay layer, 1em equates 1in. This allows non-scaling
structures. For example, the dimensions of SelectionBox handles remain
constant despite the size of the box tracking the zoom level. This
comes at a cost. All things in the overlay undergo layout calculations
when you zoom or resize the window. Consider carefully what goes in here
for performance